requirements.txt
anyio==3.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
bqplot==0.12.31
branca==0.4.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
colour==0.1.5
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.12.4
earthengine-api==0.1.292
entrypoints==0.3
ffmpeg-python==0.2.0
folium==0.12.1.post1
fonttools==4.28.3
future==0.18.2
geeadd==0.5.5
geemap==0.10.1
geetools==0.6.14
geocoder==1.38.1
geojson==2.5.0
google-api-core==1.31.4
google-api-python-client==1.12.8
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
googledrivedownloader==0.4
httplib2==0.20.2
httplib2shim==0.0.3
idna==3.3
ipyevents==2.0.1
ipyfilechooser==0.6.0
ipygee==0.0.17
ipykernel==6.6.0
ipyleaflet==0.14.0
ipynb-py-convert==0.4.6
ipython==7.30.1
ipython-genutils==0.2.0
ipytree==0.2.1
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
json5==0.9.6
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-server==1.13.1
jupyterlab==3.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
logzero==1.7.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mss==6.1.0
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
numpy==1.21.4
OWSLib==0.25.0
packaging==21.3
palettable==3.3.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
proplot==0.9.5
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyCRS==1.0.2
pygal==3.0.0
Pygments==2.10.0
pyparsing==3.0.6
pyproj==3.3.0
pyrsistent==0.18.0
pyshp==2.1.3
python-box==5.4.1
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
pywin32==302
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
ratelim==0.1.6
requests==2.26.0
rsa==4.8
sankee==0.0.7
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
sqlparse==0.4.2
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.1
traittypes==0.2.1
tzdata==2021.5
uritemplate==3.0.1
urllib3==1.26.7
voila==0.3.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
websockets==10.1
whitebox==2.0.3
whiteboxgui==0.6.0
widgetsnbextension==3.5.2
xyzservices==2021.11.0
